+++ CHANGELOG.md
@@ -12,7 +12,8 @@ This project **DOES NOT** adhere to [Semantic Versioning](http://semver.org/spec
 - implementation of `Statement#setMaxRows`
 - implementation of `Statement#setMaxFieldSize`
 - implementation of `Statement#setFetchDirection`
-- implementation of `Statement#.setPoolable`
+- implementation of `Statement#setPoolable`
+- implementation of `Statement#setLargeMaxRows`
 
 ### Removed
 - dependency on `jcip-annotations`